+++ src/main/java/org/sonar/plugins/delphi/pmd/rules/DuplicatesRule.java
@@ -6,11 +6,8 @@ import net.sourceforge.pmd.RuleContext;
 import net.sourceforge.pmd.lang.ast.Node;
 import org.sonar.plugins.delphi.antlr.ast.node.AssignmentStatementNode;
 import org.sonar.plugins.delphi.antlr.ast.node.MethodBodyNode;
+import org.sonar.plugins.delphi.antlr.ast.node.StatementListNode;
 
-/**
- * This rule adds violations when the Duplicates method, (foo.Duplicates := dupError) is called on a
- * list, but the preceding line did not first sort the list (using foo.Sorted := True)
- */
 public class DuplicatesRule extends AbstractDelphiRule {
 
   @Override
@@ -22,8 +19,7 @@ public class DuplicatesRule extends AbstractDelphiRule {
           .filter(AssignmentStatementNode.class::isInstance)
           .map(AssignmentStatementNode.class::cast)
           .filter(DuplicatesRule::isDuplicatesStatement)
-          .filter(not(DuplicatesRule::isSortedByPreviousStatement))
-          .filter(not(DuplicatesRule::isSortedByNextStatement))
+          .filter(not(DuplicatesRule::isSortedInSameBlock))
           .forEach(statement -> addViolation(data, statement));
     }
     return super.visit(methodBody, data);
@@ -43,14 +39,12 @@ public class DuplicatesRule extends AbstractDelphiRule {
     return assigneeImage.regionMatches(true, assigneeImage.length() - 11, ".Duplicates", 0, 11);
   }
 
-  private static boolean isSortedByPreviousStatement(AssignmentStatementNode duplicates) {
-    Node previous = duplicates.jjtGetParent().jjtGetChild(duplicates.jjtGetChildIndex() - 2);
-    return isSortedStatement(duplicates, previous);
-  }
-
-  private static boolean isSortedByNextStatement(AssignmentStatementNode duplicates) {
-    Node previous = duplicates.jjtGetParent().jjtGetChild(duplicates.jjtGetChildIndex() + 2);
-    return isSortedStatement(duplicates, previous);
+  private static boolean isSortedInSameBlock(AssignmentStatementNode duplicates) {
+    Node parent = duplicates.jjtGetParent();
+    return parent instanceof StatementListNode
+        && ((StatementListNode) parent)
+            .statementStream()
+            .anyMatch(statement -> isSortedStatement(duplicates, statement));
   }
 
   private static boolean isSortedStatement(AssignmentStatementNode duplicates, Node statement) {